+++ R/shift.R
@@ -47,7 +47,7 @@ weighted_shift <- function(mat) {
 }
 
 
-counts_shift_inner <- function(counts, groups) {
+counts_shift_inner <- function(counts, groups, typecast) {
     relevant <- unique(unlist(groups))
     counts <- as.matrix(counts[relevant,,drop=FALSE])
     
@@ -78,8 +78,8 @@ counts_shift_inner <- function(counts, groups) {
 
     SummarizedExperiment(
         assays=list(
-            x=realize_if_delayed(shifts),
-            weights=realize_if_delayed(weights)),
+            x=typecast(shifts),
+            weights=typecast(weights)),
         rowData=data.frame(
             per_read_var = per_read_vars,
             total_reads = totals))
@@ -110,6 +110,11 @@ counts_shift_inner <- function(counts, groups) {
 #' @param verbose 
 #' If TRUE, output some debugging and progress information.
 #'
+#' @param typecast
+#' A function to convert a matrix to a matrix-like type.
+#' Applied at the chunk level, before all chunks are \code{rbind}ed.
+#' Allows use of memory-efficient matrix representations.
+#'
 #' @return 
 #' A SummarizedExperiment object with metadata fields marking it as a weitrix.
 #' 
@@ -132,7 +137,7 @@ counts_shift_inner <- function(counts, groups) {
 #' rowData(wei)
 #'
 #' @export
-counts_shift <- function(counts, grouping, verbose=TRUE) {
+counts_shift <- function(counts, grouping, verbose=TRUE, typecast=identity) {
     assert_that(
         is.data.frame(grouping), 
         "group" %in% colnames(grouping), 
@@ -149,7 +154,7 @@ counts_shift <- function(counts, grouping, verbose=TRUE) {
         message("Calculating shifts in ",length(parts)," blocks")
 
     result <- lapply(parts, function(part) {
-        counts_shift_inner(counts, groups[part])
+        counts_shift_inner(counts, groups[part], typecast=typecast)
     })
     result <- do.call(rbind, result)
     colnames(result) <- colnames(counts)